Time: 3 ms (38.28%), Space: 8.1 MB (45.66%) - LeetHub

+++ 0050-powx-n/0050-powx-n.cpp
@@ -3,8 +3,7 @@ public:
     double myPow(double x, int n) {
         double ans = 1.0;
         long long nn = n;
-        
-        if(nn<0) nn = -1 * nn;
+        if(nn < 0) nn = -1 * nn;
         
         while(nn){
             if(nn%2 == 0){
@@ -15,7 +14,7 @@ public:
                 nn = nn - 1;
             }
         }
-        if(n<0) ans = 1/ans;
+        if(n<0) ans = (double)(1.0)/(double)ans;
         return ans;
     }
 };
